Move `constants.js` and `data-schema.json`
